Move index and admin to containers
